+++ backend/handlers/routes/domain.js
@@ -1,4 +1,5 @@
-const { Domain } = require('../../models');
+const { Domain, Story } = require('../../models');
+const { retreiveStories, addStoryToDomain } = require('../database/domain');
 
 async function retrieveById (req, res, next) {
   try {
@@ -12,6 +13,33 @@ async function retrieveById (req, res, next) {
   }
 }
 
+async function retreiveAllDomainStories (req, res, next) {
+  try {
+    const stories = await retreiveStories(req.params.domainID);
+    return res.status(200).json(stories);
+  } catch (error) {
+    next(error);
+  }
+}
+
+async function createStory (req, res, next) {
+  try {
+    let story = await Story.create(req.body);
+    const { domainID } = req.params;
+    const domain = await addStoryToDomain(story._id, domainID);
+    console.log(domain);
+    if (!domain) {
+      throw new Error('Invalid Domain ID');
+    } else {
+    }
+    return res.status(200).json(story);
+  } catch (error) {
+    next(error);
+  }
+}
+
 module.exports = {
-  retrieveById
+  retrieveById,
+  retreiveAllDomainStories,
+  createStory
 };
