+++ backend/handlers/routes/story.js
@@ -13,30 +13,30 @@ const {
   retrieveActionsFromStory
 } = require('../database/story');
 
-async function retreiveAllDomainStories (req, res, next) {
-  try {
-    const stories = await retreiveStories(req.params.domainID);
-    return res.status(200).json(stories);
-  } catch (error) {
-    next(error);
-  }
-}
+// async function retreiveAllDomainStories (req, res, next) {
+//   try {
+//     const stories = await retreiveStories(req.params.domainID);
+//     return res.status(200).json(stories);
+//   } catch (error) {
+//     next(error);
+//   }
+// }
 
-async function create (req, res, next) {
-  try {
-    let story = await Story.create(req.body);
-    const { domainID } = req.params;
-    const domain = await addStoryToDomain(story._id, domainID);
-    console.log(domain);
-    if (!domain) {
-      throw new Error('Invalid Domain ID');
-    } else {
-    }
-    return res.status(200).json(story);
-  } catch (error) {
-    next(error);
-  }
-}
+// async function create (req, res, next) {
+//   try {
+//     let story = await Story.create(req.body);
+//     const { domainID } = req.params;
+//     const domain = await addStoryToDomain(story._id, domainID);
+//     console.log(domain);
+//     if (!domain) {
+//       throw new Error('Invalid Domain ID');
+//     } else {
+//     }
+//     return res.status(200).json(story);
+//   } catch (error) {
+//     next(error);
+//   }
+// }
 
 async function retrieveStoryAndEvents (req, res, next) {
   try {
@@ -60,7 +60,7 @@ async function addIntent (req, res, next) {
   try {
     const intent = await Intent.create(req.body);
     const story = await addIntentToStory(intent._id, req.params.storyID);
-    await addIntentToDomain(intent._id, req.params.domainID);
+    await addIntentToDomain(intent._id, intent.domainID);
     return res.status(200).json(story);
   } catch (error) {
     next(error);
@@ -80,7 +80,7 @@ async function addAction (req, res, next) {
   try {
     const action = await Action.create(req.body);
     const story = await addActionToStory(action._id, req.params.storyID);
-    await addActionToDomain(action._id, req.params.domainID);
+    await addActionToDomain(action._id, action.domainID);
     return res.status(200).json(story);
   } catch (error) {
     next(error);
@@ -89,10 +89,8 @@ async function addAction (req, res, next) {
 
 async function modifyStory (req, res, next) {
   try {
-    const newStory = await Story.findByIdAndUpdate(
-      req.params.storyID,
-      req.body
-    );
+    let newStory = await Story.findByIdAndUpdate(req.params.storyID, req.body);
+    newStory = await Story.findById(newStory._id);
     return res.status(200).json(newStory);
   } catch (error) {
     next(error);
@@ -100,8 +98,6 @@ async function modifyStory (req, res, next) {
 }
 
 module.exports = {
-  retreiveAllDomainStories,
-  create,
   retrieveIntents,
   addIntent,
   retrieveActions,
